Create StableMir replacer for SMirCalls
